% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexpipe_shiny_helpers.R
\name{.gexpipe_fread_counts}
\alias{.gexpipe_fread_counts}
\title{Read tabular count files (GEO supplementary files are often messy)}
\usage{
.gexpipe_fread_counts(path, ...)
}
\description{
Read tabular count files (GEO supplementary files are often messy)
}
\keyword{internal}
