% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexpipe_shiny_helpers.R
\name{.gexpipe_geo_quiet}
\alias{.gexpipe_geo_quiet}
\title{Capture verbose GEOquery console output without suppressMessages()}
\usage{
.gexpipe_geo_quiet(expr)
}
\description{
Capture verbose GEOquery console output without suppressMessages()
}
\keyword{internal}
