% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_imports.R
\name{dummy_imports}
\alias{dummy_imports}
\title{Dummy function to satisfy R CMD check that packages in Imports are used}
\usage{
dummy_imports()
}
\description{
Dummy function to satisfy R CMD check that packages in Imports are used
}
\keyword{internal}
