% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp_platform_helpers.R
\name{gexpipe_build_batch_mod}
\alias{gexpipe_build_batch_mod}
\title{Build ComBat / removeBatchEffect model matrix (biology to preserve)}
\usage{
gexpipe_build_batch_mod(metadata)
}
\arguments{
\item{metadata}{sample metadata data.frame.}
}
\value{
model matrix
}
\description{
Build ComBat / removeBatchEffect model matrix (biology to preserve)
}
